+++ src/main/kotlin/dev/protsenko/securityLinter/utils/PsiElementGenerator.kt
@@ -2,9 +2,11 @@ package dev.protsenko.securityLinter.utils
 
 import com.intellij.docker.dockerFile.DockerFileType
 import com.intellij.docker.dockerFile.parser.psi.DockerFileFromCommand
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.fileTypes.PlainTextFileType
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiFileFactory
 import com.intellij.psi.util.childrenOfType
 
@@ -26,11 +28,13 @@ object PsiElementGenerator {
         } else {
             "FROM $image@$digest"
         }
-        val dockerFile = factory.createFileFromText(
-            DUMMY_DOCKERFILE_NAME,
-            DockerFileType.DOCKER_FILE_TYPE,
-            fromDefinition
-        )
+        val dockerFile = ApplicationManager.getApplication().runReadAction<PsiFile> {
+            factory.createFileFromText(
+                DUMMY_DOCKERFILE_NAME,
+                DockerFileType.DOCKER_FILE_TYPE,
+                fromDefinition
+            )
+        }
         return dockerFile.childrenOfType<DockerFileFromCommand>().firstOrNull()
     }